requirements.txt
# Marauder CV Pipeline - Complete Requirements
# Python 3.10+

# Core Deep Learning
torch>=2.0.0
torchvision>=0.15.0
ultralytics>=8.0.196  # YOLOv8
# ultralytics supports YOLOv11 in versions >= 8.3.0

# Computer Vision
opencv-python>=4.8.0
opencv-contrib-python>=4.8.0
Pillow>=10.0.0
albumentations>=1.3.1
imgaug>=0.4.0
kornia>=0.7.0

# Self-Supervised Learning
lightly>=1.4.0  # For MoCo V3

# Object Tracking
supervision>=0.16.0  # Includes ByteTrack

# Data Handling
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.11.0
scikit-learn>=1.3.0
scikit-image>=0.21.0

# Storage and I/O
boto3>=1.28.0  # S3/Spaces storage
botocore>=1.31.0
h5py>=3.9.0
pyarrow>=12.0.0

# API and Web
requests>=2.31.0
urllib3>=2.0.0
aiohttp>=3.8.0
httpx>=0.24.0

# Configuration
pyyaml>=6.0
omegaconf>=2.3.0
python-dotenv>=1.0.0

# Logging and Monitoring
wandb>=0.15.0  # Weights & Biases
tensorboard>=2.13.0
tqdm>=4.66.0
rich>=13.5.0
loguru>=0.7.0

# Metrics and Evaluation
torchmetrics>=1.0.0
pycocotools>=2.0.6

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# Utilities
click>=8.1.0
colorama>=0.4.6
psutil>=5.9.0
py-cpuinfo>=9.0.0

# TensorRT (for Jetson Nano deployment)
# Note: TensorRT must be installed separately on Jetson
# onnx>=1.14.0
# onnxruntime-gpu>=1.15.0

# Calibration
netcal>=1.3.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Code Quality
black>=23.7.0
flake8>=6.1.0
isort>=5.12.0

# GCP Deployment
google-cloud-storage>=2.10.0
google-cloud-aiplatform>=1.35.0

# Version Pinning
setuptools>=68.0.0
wheel>=0.41.0
